# Some more Parquet writer performance improvements