--- a/src/common/types/hash.cpp
+++ b/src/common/types/hash.cpp
@@ -78,6 +78,35 @@ hash_t Hash(const char *str) {
 
 template <>
 hash_t Hash(string_t val) {
+	// If the string is inlined, we can do a branchless hash
+	if (val.IsInlined()) {
+		// This seed slightly improves bit distribution, taken from here:
+		// https://github.com/martinus/robin-hood-hashing/blob/3.11.5/LICENSE
+		// MIT License Copyright (c) 2018-2021 Martin Ankerl
+		hash_t h = 0xe17a1465U ^ (val.GetSize() * 0xc6a4a7935bd1e995U);
+
+		// Hash/combine the first 8-byte block
+		h ^= Load<hash_t>(const_data_ptr_cast(val.GetPrefix()));
+		h *= 0xd6e8feb86659fd93U;
+
+		// Load remaining 4 bytes
+		hash_t hr = 0;
+		memcpy(&hr, const_data_ptr_cast(val.GetPrefix()) + sizeof(hash_t), 4U);
+
+		// Process the remainder the same an 8-byte block
+		// This operation is a NOP if the string is <= 8 bytes
+		const bool not_a_nop = val.GetSize() > sizeof(hash_t);
+		h ^= hr;
+		h *= 0xd6e8feb86659fd93U * not_a_nop + (1 - not_a_nop);
+
+		// This is just an optimization. It should not change the result
+		// This property is important for verification (e.g., DUCKDB_DEBUG_NO_INLINE)
+		// We achieved this with the NOP trick above (and in HashBytes)
+		h = Hash(h);
+		D_ASSERT(h == Hash(val.GetData(), val.GetSize()));
+
+		return h;
+	}
 	return Hash(val.GetData(), val.GetSize());
 }
 
@@ -98,22 +127,18 @@ hash_t HashBytes(const_data_ptr_t ptr, const idx_t len) noexcept {
 		h *= 0xd6e8feb86659fd93U;
 	}
 
-	// Load and process remaining (<8) bytes
+	// Load remaining (<8) bytes
 	hash_t hr = 0;
 	memcpy(&hr, ptr, len & 7U);
-	hr *= 0xd6e8feb86659fd93U;
-	hr ^= h >> 32;
 
-	// XOR with hash
+	// Process the remainder same as an 8-byte block
+	// This operation is a NOP if the number of remaining bytes is 0
+	const bool not_a_nop = len & 7U;
 	h ^= hr;
+	h *= 0xd6e8feb86659fd93U * not_a_nop + (1 - not_a_nop);
 
 	// Finalize
-	h *= 0xd6e8feb86659fd93U;
-	h ^= h >> 32;
-
-	return h;
-
-	// return Hash(h);
+	return Hash(h);
 }
 
 hash_t Hash(const char *val, size_t size) {