… test/sql/function/generic/hash_func.test